+++ pkg/ddc/thin/referencedataset/engine.go
@@ -1,4 +1,4 @@
-package virtualdataset
+package referencedataset
 
 import (
 	"context"
