+++ pkg/ddc/thin/referencedataset/sync.go
@@ -1,4 +1,4 @@
-package virtualdataset
+package referencedataset
 
 import (
 	"context"
